feat: Lots of little and major things.
Implemented all context menu item functionality (except opacity).

Implemented sidebar (though inert for now).

Implemented a shared class for handling dragging, and done some
mostly successfull changes to the cursor relative to its position
(though with the caveat that you'll have to move the cursor to
update it in some cases).

Created an alternative textInputAlt method that takes a Text shape
instead of just a String, with the future plans to use that to make
it easier to both apply and get the style state of the Text shape."